Move the type to atl.py
